Move unlocking back into liminal module
This makes DeviceSet a pure data structure.
